# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenG2P Approval Workflow Engine",
    "description": "Generic, configurable multi-stage approval workflow engine for OpenG2P modules. Caller services post artifacts; AWE resolves stages and approvers, then notifies the caller via signed webhook callbacks when state changes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/awe/delegations": {
      "get": {
        "tags": [
          "delegations"
        ],
        "summary": "List delegations \u2014 filter by user",
        "operationId": "list_delegations_v1_awe_delegations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DelegationOut"
                  },
                  "title": "Response List Delegations V1 Awe Delegations Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks `AWE_VIEWER` or `AWE_ADMIN` role.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_VIEWER or AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "delegations"
        ],
        "summary": "Create a delegation window (admin only)",
        "operationId": "create_delegation_v1_awe_delegations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DelegationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DelegationOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/delegations/{delegation_id}": {
      "delete": {
        "tags": [
          "delegations"
        ],
        "summary": "Delete a delegation (admin only)",
        "operationId": "delete_delegation_v1_awe_delegations__delegation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "delegation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Delegation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No delegation with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-004",
                      "message": "Delegation not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "components": {
    "schemas": {
...
      "DelegationCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "minLength": 1,
            "title": "User Id",
            "examples": [
              "u-alice"
            ]
          },
          "delegate_to": {
            "type": "string",
            "minLength": 1,
            "title": "Delegate To",
            "examples": [
              "u-bob"
            ]
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "delegate_to",
          "starts_at",
          "ends_at"
        ],
        "title": "DelegationCreate"
      },
      "DelegationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "delegate_to": {
            "type": "string",
            "title": "Delegate To"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "delegate_to",
          "starts_at",
          "ends_at",
          "created_at",
          "updated_at"
        ],
        "title": "DelegationOut"
      },
...
      "ErrorDetail": {
        "properties": {
          "errorCode": {
            "type": "string",
            "title": "Errorcode",
            "description": "OpenG2P-assigned error code. AWE catalog: `AWE-001` (policy not found), `AWE-002` (policy conflict / version clash), `AWE-003` (request not found), `AWE-004` (task not found), `AWE-005` (service not ready \u2014 startup incomplete), `AWE-006` (database health check failed), `AWE-007` (invalid state transition), `AWE-008` (unauthorized / forbidden), `AWE-009` (idempotency key conflict with different payload), `AWE-010` (validation \u2014 bad policy definition).",
            "examples": [
              "AWE-003"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Approval request not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": "openg2p.awe",
            "examples": [
              "openg2p.awe"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "errors"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}